Moved workflow option dropdown to its own component
