+++ client-react/src/pages/app/deployment-center/DeploymentCenter.types.ts
@@ -35,7 +35,6 @@ export enum DeploymentStatus {
 
 export enum WorkflowOption {
   None = 'none',
-  Loading = 'loading',
   Overwrite = 'overwrite',
   Add = 'add',
   UseAvailableWorkflowConfigs = 'useAvailableWorkflowConfigs',
