+++ client-react/src/pages/app/deployment-center/code/DeploymentCenterCodeSettings.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { DeploymentCenterFieldProps, DeploymentCenterCodeFormData } from '../DeploymentCenter.types';
+import { DeploymentCenterFieldProps, DeploymentCenterCodeFormData, WorkflowOption } from '../DeploymentCenter.types';
 import DeploymentCenterGitHubDataLoader from '../github-provider/DeploymentCenterGitHubDataLoader';
 import { ScmType, BuildProvider } from '../../../../models/site/config';
 import DeploymentCenterCodeBuild from './DeploymentCenterCodeBuild';
@@ -8,6 +8,7 @@ import DeploymentCenterGitHubReadOnly from '../github-provider/DeploymentCenterG
 import DeploymentCenterCodeBuildReadOnly from './DeploymentCenterCodeBuildReadOnly';
 import DeploymentCenterGitHubWorkflowConfig from '../github-provider/DeploymentCenterGitHubWorkflowConfig';
 import DeploymentCenterCodeSourceAndBuild from './DeploymentCenterCodeSourceAndBuild';
+import DeploymentCenterGitHubWorkflowConfigSelector from '../github-provider/DeploymentCenterGitHubWorkflowConfigSelector';
 
 const DeploymentCenterCodeSettings: React.FC<DeploymentCenterFieldProps<DeploymentCenterCodeFormData>> = props => {
   const { formProps } = props;
@@ -16,6 +17,10 @@ const DeploymentCenterCodeSettings: React.FC<DeploymentCenterFieldProps<Deployme
   const isGitHubSource = formProps && formProps.values.sourceProvider === ScmType.GitHub;
   const isGitHubActionsBuild = formProps && formProps.values.buildProvider === BuildProvider.GitHubAction;
   const isDeploymentSetup = deploymentCenterContext.siteConfig && deploymentCenterContext.siteConfig.properties.scmType !== ScmType.None;
+  const isUsingExistingOrAvailableWorkflowConfig =
+    formProps &&
+    (formProps.values.workflowOption === WorkflowOption.UseExistingWorkflowConfig ||
+      formProps.values.workflowOption === WorkflowOption.UseAvailableWorkflowConfigs);
 
   const disconnectCallback = () => {
     throw Error('not implemented');
@@ -31,8 +36,13 @@ const DeploymentCenterCodeSettings: React.FC<DeploymentCenterFieldProps<Deployme
       ) : (
         <>
           <DeploymentCenterCodeSourceAndBuild formProps={formProps} />
-          {isGitHubSource && <DeploymentCenterGitHubDataLoader formProps={formProps} />}
-          {isGitHubActionsBuild && <DeploymentCenterCodeBuild formProps={formProps} />}
+          {isGitHubSource && (
+            <>
+              <DeploymentCenterGitHubDataLoader formProps={formProps} />
+              {isGitHubActionsBuild && <DeploymentCenterGitHubWorkflowConfigSelector formProps={formProps} />}
+            </>
+          )}
+          {isGitHubActionsBuild && !isUsingExistingOrAvailableWorkflowConfig && <DeploymentCenterCodeBuild formProps={formProps} />}
           <DeploymentCenterGitHubWorkflowConfig formProps={formProps} />
         </>
       )}
